requirements.txt
fastapi==0.115.5
starlette==0.41.3
uvicorn==0.30.6
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
